// apps/pc/src/generated/models/Pl_requestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_requestsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_requestsstatecode = keyof typeof Pl_requestsstatecode;
export const Pl_requestsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_requestsstatuscode = keyof typeof Pl_requestsstatuscode;

export interface Pl_requestsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_activeapprovalkey?: string;
  pl_approvalpolicyversion?: string;
  "pl_ApproverTeamLookup@odata.bind"?: string;
  pl_cancellationreason?: string;
  "pl_ContractLookup@odata.bind"?: string;
  pl_name?: string;
  "pl_PartnerLookup@odata.bind"?: string;
  pl_requestedat?: string;
  pl_requestid: string;
  "pl_RequestingUserLookup@odata.bind"?: string;
  pl_requestkey?: string;
  pl_requeststatuscode?: string;
  pl_requesttypecode?: string;
  statecode: Pl_requestsstatecode;
  statuscode?: Pl_requestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_requests extends Pl_requestsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_approverteamlookupname?: string;
  pl_approverteamlookupyominame?: string;
  pl_contractlookupname?: string;
  pl_partnerlookupname?: string;
  pl_requestinguserlookupname?: string;
  pl_requestinguserlookupyominame?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_approverteamlookup?: object;
  _pl_approverteamlookup_value?: string;
  pl_contractlookup?: object;
  _pl_contractlookup_value?: string;
  pl_partnerlookup?: object;
  _pl_partnerlookup_value?: string;
  pl_requestinguserlookup?: object;
  _pl_requestinguserlookup_value?: string;
}