// apps/pc/src/generated/models/Pl_submissionversionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_submissionversionsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_submissionversionsstatecode = keyof typeof Pl_submissionversionsstatecode;
export const Pl_submissionversionsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_submissionversionsstatuscode = keyof typeof Pl_submissionversionsstatuscode;

export interface Pl_submissionversionsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_approvaltitle?: string;
  pl_changesetjson?: string;
  pl_changesummary?: string;
  pl_fixedat?: string;
  pl_name?: string;
  pl_policyversion?: string;
  "pl_RequestLookup@odata.bind"?: string;
  pl_rowversiontoken?: string;
  pl_submissionstatuscode?: string;
  pl_submissionversionid: string;
  pl_submittedat?: string;
  "pl_SubmittedByLookup@odata.bind"?: string;
  pl_versionnumber?: number;
  statecode: Pl_submissionversionsstatecode;
  statuscode?: Pl_submissionversionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_submissionversions extends Pl_submissionversionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_requestlookupname?: string;
  pl_submittedbylookupname?: string;
  pl_submittedbylookupyominame?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_requestlookup?: object;
  _pl_requestlookup_value?: string;
  pl_submittedbylookup?: object;
  _pl_submittedbylookup_value?: string;
}